finished adding in ac and DCpow function
I fixed the distortion measurement in the tran function and wrote up the ac and DCpow function. I think it might be better to move the Monte Carlo function to the GA submodule since the measure of the robustness of the model is likely going to be something like the variance of the fitness scores. I tried to make this submodule as fast as possible but it looks like the main bottleneck by a wide margin is simulating the circuit and running the commands on LTspice. For the transient function, it takes a total of roughly .6335 seconds and it takes .6281 seconds to run the simulation. For the ac function, it takes about .1040 seconds to run and .01021 seconds to run the simulation. And finally for the DCpow function, it takes .05580 seconds to run and .05442 seconds to simulate. So in total to calculate these measurements for a circuit it probably would take around .7437 seconds. This might be an issue in the future because as the circuits get more complicated, it could take longer to simulate them. And we also have to do multiple repetitions to measure the sensitivity of the circuit so if we do maybe 5 repetitions, calculating fitness score for a single circuit could take at least 3.7 seconds. I think I might make the distortion measurement optional; for amplifiers it's not really an important metric I think. That would make this take around .0636 seconds which is significantly faster and puts us at a total of .8691 seconds to calculate the fitness score for a single circuit which is a lot better.